Offline Weekend
I was offline this weekend.  There's a lot of changes.  The most notable is that I did the assignment.  It's difficult to break that down into steps, but let's just say that I gazed into the void that is three.js and I would only hope someday to worship at its dark altar.  Cool beans, now, bye bye.